copilot-adk-app/backend/pyproject.toml
[project]
name = "copilot-adk-backend"
version = "0.1.0"
description = "ADK Python agent with Postgres sessions and auth for CopilotKit frontend"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "google-adk>=1.0.0",
    "ag-ui-adk>=0.3.0",
    "asyncpg>=0.28.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
